Keep generated table Doc client-safe
The documented way to type function returns is `notes.Doc` from
`_generated/tables/notes`. That wrapper bound `Table.make` from
`@confect/server`, so any client that imported `_generated/refs`
also loaded the server package and `convex/server` via
`tableDefinition`.

`Table.make`, `Fields`, and `Doc` now live in `@confect/core`.
`tableDefinition` stays a server helper used only by generated
`convexSchema.ts`. Codegen's spec-graph check no longer exempts
`tables/` and `_generated/`, so a spec that uses `notes.Doc` cannot
reach `@confect/server`.
