+++ apps/docs/getting-started/quickstart.mdx
@@ -43,7 +43,7 @@ bun add convex @confect/core @confect/server @confect/cli @confect/react
     Define a table — one file per table under `confect/tables/`, where the **filename is the table name** (so filenames must be valid JS identifiers and may not start with `_`). Each file **must default-export** a `Table`; codegen reads `module.default` and applies the filename as the table name. From other modules (e.g. specs), default-import the codegen-emitted wrapper at `confect/_generated/tables/<name>` to reach the table's derived `Schema`s (`notes.Doc`, `notes.Fields`).
 
     ```ts confect/tables/notes.ts
-    import { Table } from "@confect/server";
+    import { Table } from "@confect/core";
     import * as Schema from "effect/Schema";
 
     export default Table.make(() =>
