+++ packages/server/test/mock-backend/fixtures/confect/_generated/convexSchema.ts
@@ -1,4 +1,5 @@
 import { defineSchema as $defineSchema } from "convex/server";
+import { Table as $Table } from "@confect/server";
 
 import events from "./tables/events";
 import notes from "./tables/notes";
@@ -6,8 +7,8 @@ import tags from "./tables/tags";
 import users from "./tables/users";
 
 export default $defineSchema({
-  events: events.tableDefinition,
-  notes: notes.tableDefinition,
-  tags: tags.tableDefinition,
-  users: users.tableDefinition,
+  events: $Table.tableDefinition(events),
+  notes: $Table.tableDefinition(notes),
+  tags: $Table.tableDefinition(tags),
+  users: $Table.tableDefinition(users),
 });
