+++ packages/server/test/mock-backend/fixtures/confect/tables/users.ts
@@ -1,4 +1,4 @@
-import { Table } from "@confect/server";
+import { Table } from "@confect/core";
 import * as Schema from "effect/Schema";
 
 export default Table.make(() =>
